+++ src/ui/views/ProtectPlayers.tsx
@@ -15,7 +15,8 @@ import { PHASE } from "../../common";
 const PlayerList = ({
 	challengeNoRatings,
 	updateProtectedPids,
-	numProtectedPlayers,
+	numPerTeam,
+	numRemaining,
 	players,
 	protectedPids,
 	stats,
@@ -23,7 +24,8 @@ const PlayerList = ({
 	upcomingFreeAgentsText,
 }: Pick<View<"protectPlayers">, "challengeNoRatings" | "players" | "stats"> & {
 	updateProtectedPids: (newProtectedPids: number[]) => void;
-	numProtectedPlayers: number;
+	numPerTeam: number;
+	numRemaining: number;
 	protectedPids: number[];
 	tid: number;
 	upcomingFreeAgentsText: React.ReactNode;
@@ -40,8 +42,6 @@ const PlayerList = ({
 		...stats.map(stat => `stat:${stat}`),
 		"Acquired",
 	);
-
-	const numRemaining = numProtectedPlayers - protectedPids.length;
 
 	const rows = players.map(p => {
 		return {
@@ -95,7 +95,11 @@ const PlayerList = ({
 				are not protected may be selected in the expansion draft.
 			</p>
 			{upcomingFreeAgentsText}
-			<p>Protections remaining: {numRemaining}</p>
+			<p>
+				Protections remaining: {numRemaining}
+				<br />
+				Max number of selected players per existing team: {numPerTeam ?? "???"}
+			</p>
 			<div className="btn-group mb-3">
 				<button
 					type="button"
@@ -147,8 +151,12 @@ const ProtectPlayers = ({
 
 	const protectedPids = expansionDraft.protectedPids[userTid] || [];
 
-	const numRemaining =
-		expansionDraft.numProtectedPlayers - protectedPids.length;
+	const maxNumCanProtext = Math.min(
+		expansionDraft.numProtectedPlayers,
+		players.length - (expansionDraft.numPerTeam ?? 0),
+	);
+
+	const numRemaining = maxNumCanProtext - protectedPids.length;
 
 	const handleSubmit = async (event: FormEvent) => {
 		event.preventDefault();
@@ -242,6 +250,10 @@ const ProtectPlayers = ({
 							Your team is an expansion team, so you have no players to protect.
 						</p>
 						{upcomingFreeAgentsText}
+						<p>
+							Max number of selected players per existing team:{" "}
+							{expansionDraft.numPerTeam ?? "???"}
+						</p>
 					</>
 				) : expansionDraft.numProtectedPlayers <= 0 ? (
 					<>
@@ -254,7 +266,8 @@ const ProtectPlayers = ({
 					<PlayerList
 						challengeNoRatings={challengeNoRatings}
 						updateProtectedPids={updateProtectedPids}
-						numProtectedPlayers={expansionDraft.numProtectedPlayers}
+						numPerTeam={expansionDraft.numPerTeam}
+						numRemaining={numRemaining}
 						players={players}
 						protectedPids={protectedPids}
 						stats={stats}
@@ -266,7 +279,7 @@ const ProtectPlayers = ({
 				<button
 					type="submit"
 					className="btn btn-primary"
-					disabled={numRemaining < 0 || saving}
+					disabled={(!expansionTeam && numRemaining < 0) || saving}
 				>
 					Start Expansion Draft
 				</button>