+++ src/worker/core/draft/runPicks.ts
@@ -43,6 +43,22 @@ const runPicks = async (
 		playersAll = (
 			await idb.cache.players.indexGetAll("playersByTid", [0, Infinity])
 		).filter(p => expansionDraft.availablePids.includes(p.pid));
+
+		if (expansionDraft.numPerTeam !== undefined) {
+			// Keep logic in sync with draft.ts
+			const tidsOverLimit: number[] = [];
+			for (const [tidString, numPerTeam] of Object.entries(
+				expansionDraft.numPerTeamDrafted,
+			)) {
+				if (numPerTeam >= expansionDraft.numPerTeam) {
+					const tid = parseInt(tidString);
+					tidsOverLimit.push(tid);
+				}
+			}
+			if (tidsOverLimit.length > 0) {
+				playersAll = playersAll.filter(p => !tidsOverLimit.includes(p.tid));
+			}
+		}
 	} else {
 		playersAll = (
 			await idb.cache.players.indexGetAll("playersByDraftYearRetiredYear", [